Move integration tests to integration tests dir
