+++ examples/test-dashboard.json
@@ -206,7 +206,7 @@
             "refId": "A"
           }
         ],
-        "title": "Liam's Panel Title",
+        "title": "Panel Title",
         "type": "timeseries"
       }
     ],
